PEAR-1860 - consolidate api for endpoint slices, remove some redundant slices
