+++ packages/core/src/features/gdcapi/gdcapi.ts
@@ -4,9 +4,17 @@ import { GqlOperation } from "./filters";
 import "isomorphic-fetch";
 import { GDC_API, GDC_APP_API_AUTH } from "../../constants";
 import { coreCreateApi } from "src/coreCreateApi";
-import { caseSummaryDefaults } from "../cases/types";
 import Queue from "queue";
 import md5 from "blueimp-md5";
+import {
+  CaseDefaults,
+  ProjectDefaults,
+  AnnotationDefaults,
+  SSMSDefaults,
+  FileDefaults,
+  HistoryDefaults,
+  GenesDefaults,
+} from "./types";
 
 export type UnknownJson = Record<string, unknown>;
 export interface GdcApiResponse<H = UnknownJson> {
@@ -177,317 +185,11 @@ export const fetchGdcCasesMapping = async (): Promise<GdcApiMapping> => {
   throw await buildFetchError(res);
 };
 
-export interface CaseDefaults {
-  readonly case_id: string;
-  readonly submitter_id: string;
-
-  readonly consent_type?: string;
-  readonly days_to_consent?: number;
-  readonly days_to_lost_to_followup?: number;
-  readonly disease_type?: string;
-  readonly index_date?: string;
-  readonly lost_to_followup?: string;
-  readonly primary_site?: string;
-
-  readonly aliquot_ids?: ReadonlyArray<string>;
-  readonly analyte_ids?: ReadonlyArray<string>;
-  readonly diagnosis_ids?: ReadonlyArray<string>;
-  readonly portion_ids?: ReadonlyArray<string>;
-  readonly sample_ids?: ReadonlyArray<string>;
-  readonly slide_ids?: ReadonlyArray<string>;
-  readonly submitter_aliquot_ids?: ReadonlyArray<string>;
-  readonly submitter_analyte_ids?: ReadonlyArray<string>;
-  readonly submitter_diagnosis_ids?: ReadonlyArray<string>;
-  readonly submitter_portion_ids?: ReadonlyArray<string>;
-  readonly submitter_sample_ids?: ReadonlyArray<string>;
-  readonly submitter_slide_ids?: ReadonlyArray<string>;
-
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  [key: string]: any;
-}
-
 export const fetchGdcCases = async (
   request?: GdcApiRequest,
-): Promise<GdcApiResponse<caseSummaryDefaults>> => {
+): Promise<GdcApiResponse<CaseDefaults>> => {
   return fetchGdcEntities("cases", request);
 };
-
-export interface ProjectDefaults {
-  readonly dbgap_accession_number: string;
-  readonly disease_type: Array<string>;
-  readonly name: string;
-  readonly primary_site: Array<string>;
-  readonly project_id: string;
-  readonly summary?: {
-    readonly case_count: number;
-    readonly file_count: number;
-    readonly file_size: number;
-    readonly data_categories?: Array<{
-      readonly case_count: number;
-      readonly data_category: string;
-      readonly file_count: number;
-    }>;
-    readonly experimental_strategies?: Array<{
-      readonly case_count: number;
-      readonly experimental_strategy: string;
-      readonly file_count: number;
-    }>;
-  };
-  readonly program?: {
-    readonly dbgap_accession_number: string;
-    readonly name: string;
-    readonly program_id: string;
-  };
-}
-
-export interface AnnotationDefaults {
-  readonly id: string;
-  readonly entity_submitter_id: string;
-  readonly notes: string;
-  readonly submitter_id: string;
-  readonly classification: string;
-  readonly entity_id: string;
-  readonly created_datetime: string;
-  readonly annotation_id: string;
-  readonly entity_type: string;
-  readonly updated_datetime: string;
-  readonly case_id: string;
-  readonly state: string;
-  readonly category: string;
-  readonly case_submitter_id: string;
-  readonly status: string;
-  readonly project?: {
-    readonly primary_site: Array<string>;
-    readonly dbgap_accession_number: string;
-    readonly project_id: string;
-    readonly disease_type: Array<string>;
-    readonly name: string;
-    readonly releasable: boolean;
-    readonly state: string;
-    readonly released: boolean;
-    readonly program?: {
-      readonly name: string;
-      readonly program_id: string;
-      readonly dbgap_accession_number: string;
-    };
-  };
-}
-
-interface transcript {
-  aa_change: string;
-  aa_end: number;
-  aa_start: number;
-  annotation: {
-    ccds: string;
-    dbsnp_rs: string;
-    existing_variation: string;
-    hgvsc: string;
-    polyphen_impact: string;
-    polyphen_score: number;
-    pubmed: string;
-    sift_impact: string;
-    sift_score: number;
-    transcript_id: string;
-    vep_impact: string;
-  };
-  is_canonical: boolean;
-  transcript_id: string;
-}
-export interface SSMSDefaults {
-  id: string;
-  consequence: Array<{ transcript: transcript }>;
-  clinical_annotations?: {
-    civic: {
-      variant_id: string;
-    };
-  };
-  reference_allele: string;
-  ncbi_build: string;
-  cosmic_id: Array<string>;
-  mutation_subtype: string;
-  chromosome: string;
-  genomic_dna_change: string;
-}
-export interface HistoryDefaults {
-  readonly uuid: string;
-  readonly version: string;
-  readonly file_change: string;
-  readonly release_date: string;
-  readonly data_release: string;
-}
-
-export interface FileDefaults {
-  readonly id?: string;
-  readonly submitter_id: string;
-  readonly access: string;
-  readonly acl: ReadonlyArray<string>;
-  readonly created_datetime: string;
-  readonly updated_datetime: string;
-  readonly data_category: string;
-  readonly data_format: string;
-  readonly data_release?: string;
-  readonly data_type: string;
-  readonly file_id: string;
-  readonly file_name: string;
-  readonly file_size: number;
-  readonly md5sum: string;
-  readonly platform?: string;
-  readonly state: string;
-  readonly type: string;
-  readonly version?: string;
-  readonly experimental_strategy?: string;
-  readonly annotations?: ReadonlyArray<{
-    readonly annotation_id: string;
-    readonly category: string;
-    readonly classification: string;
-    readonly created_datetime: string;
-    readonly entity_id: string;
-    readonly entity_submitter_id: string;
-    readonly entity_type: string;
-    readonly notes: string;
-    readonly state: string;
-    readonly status: string;
-    readonly updated_datetime: string;
-  }>;
-  readonly cases?: ReadonlyArray<{
-    readonly case_id: string;
-    readonly submitter_id: string;
-    readonly annotations?: ReadonlyArray<{
-      readonly annotation_id: string;
-    }>;
-    readonly project?: {
-      readonly dbgap_accession_number?: string;
-      readonly disease_type: string;
-      readonly name: string;
-      readonly primary_site: string;
-      readonly project_id: string;
-      readonly releasable: boolean;
-      readonly released: boolean;
-      readonly state: string;
-    };
-    readonly samples?: ReadonlyArray<{
-      readonly sample_id: string;
-      readonly sample_type: string;
-      readonly tissue_type: string;
-      readonly submitter_id: string;
-      readonly portions?: ReadonlyArray<{
-        readonly submitter_id: string;
-        readonly analytes?: ReadonlyArray<{
-          readonly analyte_id: string;
-          readonly analyte_type: string;
-          readonly submitter_id: string;
-          readonly aliquots?: ReadonlyArray<{
-            readonly aliquot_id: string;
-            readonly submitter_id: string;
-          }>;
-        }>;
-        readonly slides?: ReadonlyArray<{
-          readonly created_datetime: string | null;
-          readonly number_proliferating_cells: number | null;
-          readonly percent_eosinophil_infiltration: number | null;
-          readonly percent_granulocyte_infiltration: number | null;
-          readonly percent_inflam_infiltration: number | null;
-          readonly percent_lymphocyte_infiltration: number | null;
-          readonly percent_monocyte_infiltration: number | null;
-          readonly percent_neutrophil_infiltration: number | null;
-          readonly percent_necrosis: number | null;
-          readonly percent_normal_cells: number | null;
-          readonly percent_stromal_cells: number | null;
-          readonly percent_tumor_cells: number | null;
-          readonly percent_tumor_nuclei: number | null;
-          readonly section_location: string | null;
-          readonly slide_id: string | null;
-          readonly state: string | null;
-          readonly submitter_id: string | null;
-          readonly updated_datetime: string | null;
-        }>;
-      }>;
-    }>;
-  }>;
-  readonly associated_entities?: ReadonlyArray<{
-    readonly entity_submitter_id: string;
-    readonly entity_type: string;
-    readonly case_id: string;
-    readonly entity_id: string;
-  }>;
-  readonly analysis?: {
-    readonly workflow_type: string;
-    readonly updated_datetime: string;
-    readonly input_files?: ReadonlyArray<{
-      readonly file_name: string;
-      readonly data_category: string;
-      readonly data_type: string;
-      readonly data_format: string;
-      readonly file_size: number;
-      readonly file_id: string;
-      readonly state: string;
-      readonly submitter_id: string;
-      readonly access: string;
-      readonly created_datetime: string;
-      readonly updated_datetime: string;
-      readonly md5sum: string;
-    }>;
-    readonly metadata?: {
-      readonly read_groups: Array<{
-        readonly read_group_id: string;
-        readonly is_paired_end: boolean;
-        readonly read_length: number;
-        readonly library_name: string;
-        readonly sequencing_center: string;
-        readonly sequencing_date: string;
-      }>;
-    };
-  };
-  readonly downstream_analyses?: ReadonlyArray<{
-    readonly workflow_type: string;
-    readonly output_files?: ReadonlyArray<{
-      readonly file_name: string;
-      readonly data_category: string;
-      readonly data_type: string;
-      readonly data_format: string;
-      readonly file_size: number;
-      readonly file_id: string;
-      readonly state: string;
-      readonly submitter_id: string;
-      readonly access: string;
-      readonly created_datetime: string;
-      readonly updated_datetime: string;
-      readonly md5sum: string;
-    }>;
-  }>;
-  readonly index_files?: ReadonlyArray<{
-    readonly submitter_id: string;
-    readonly created_datetime: string;
-    readonly updated_datetime: string;
-    readonly data_category: string;
-    readonly data_format: string;
-    readonly data_type: string;
-    readonly file_id: string;
-    readonly file_name: string;
-    readonly file_size: number;
-    readonly md5sum: string;
-    readonly state: string;
-  }>;
-}
-
-export interface GenesDefaults {
-  readonly biotype: string;
-  readonly symbol: string;
-  readonly cytoband: ReadonlyArray<string>;
-  readonly synonyms: ReadonlyArray<string>;
-  readonly description: string;
-  readonly canonical_transcript_id: string;
-  readonly canonical_transcript_length: number;
-  readonly canonical_transcript_length_cds: number;
-  readonly canonical_transcript_length_genomic: string;
-  readonly gene_chromosome: string;
-  readonly gene_end: string;
-  readonly gene_id: string;
-  readonly gene_start: number;
-  readonly gene_strand: string;
-  readonly is_cancer_gene_census: boolean;
-  readonly name: string;
-}
 
 /**
  * Fetches GDC projects from the GDC REST API
@@ -668,17 +370,18 @@ export const getGdcInstance = async <T>(
  * - add auth header
  */
 
+export interface EndpointRequestProps {
+  readonly request: GdcApiRequest;
+  readonly fetchAll?: boolean;
+}
+
 export const endpointSlice = coreCreateApi({
   reducerPath: "entities",
   baseQuery: async ({
     request,
     endpoint,
-    fetchAll = true,
-  }: {
-    request: GdcApiRequest;
-    endpoint: gdcEndpoint;
-    fetchAll?: boolean;
-  }) => {
+    fetchAll,
+  }: EndpointRequestProps & { endpoint: string }) => {
     let results;
 
     try {
@@ -691,56 +394,49 @@ export const endpointSlice = coreCreateApi({
   },
   refetchOnReconnect: true,
   endpoints: (builder) => ({
-    getGenes: builder.query({
-      query: ({
-        request,
-        fetchAll,
-      }: {
-        request: GdcApiRequest;
-        fetchAll?: boolean;
-      }) => ({
+    getGenes: builder.query<GdcApiData<GenesDefaults>, EndpointRequestProps>({
+      query: ({ request, fetchAll = false }) => ({
         request,
         endpoint: "genes",
         fetchAll,
       }),
-      transformResponse: (response) => response.data.hits,
+      transformResponse: (response: GdcApiResponse<GenesDefaults>) =>
+        response.data,
     }),
-    getCases: builder.query({
-      query: (request: GdcApiRequest) => ({
+    getCases: builder.query<GdcApiData<CaseDefaults>, EndpointRequestProps>({
+      query: ({ request, fetchAll = false }) => ({
         request,
         endpoint: "cases",
-      }),
-      transformResponse: (response) => response.data.hits,
-    }),
-    getSsms: builder.query({
-      query: ({
-        request,
         fetchAll,
-      }: {
-        request: GdcApiRequest;
-        fetchAll?: boolean;
-      }) => ({
+      }),
+      transformResponse: (response: GdcApiResponse<CaseDefaults>) =>
+        response.data,
+    }),
+    getSsms: builder.query<GdcApiData<SSMSDefaults>, EndpointRequestProps>({
+      query: ({ request, fetchAll = false }) => ({
         request,
         endpoint: "ssms",
         fetchAll,
       }),
-      transformResponse: (response) => response.data.hits,
+      transformResponse: (response: GdcApiResponse<SSMSDefaults>) =>
+        response.data,
     }),
     getCaseSsms: builder.query({
-      query: (request: GdcApiRequest) => ({
+      query: ({ request, fetchAll = false }) => ({
         request,
         endpoint: "case_ssms",
+        fetchAll,
       }),
-      transformResponse: (response) => response.data.hits,
+      transformResponse: (response) => response.data,
     }),
     getAnnotations: builder.query<
       GdcApiData<AnnotationDefaults>,
-      GdcApiRequest
+      EndpointRequestProps
     >({
-      query: (request: GdcApiRequest) => ({
+      query: ({ request, fetchAll = false }) => ({
         request,
         endpoint: "annotations",
-        fetchAll: false,
+        fetchAll,
       }),
       transformResponse: (response: GdcApiResponse<AnnotationDefaults>) =>
         response.data,
