+++ packages/core/src/index.ts
@@ -7,6 +7,7 @@ export * from "./constants";
 export * from "./features/gdcapi/gdcapi";
 export * from "./features/gdcapi/filters";
 export * from "./features/gdcapi/gdcgraphql";
+export * from "./features/gdcapi/types";
 export * from "./features/imageDetails/imageDetailsSlice";
 export * from "./features/imageDetails/imageViewer";
 export * from "./features/facets/facetSlice";
@@ -25,7 +26,6 @@ export * from "./features/projects/projectsSlice";
 export * from "./features/projects/projectsPrimarySiteSlice";
 export * from "./features/history/historySlice";
 export * from "./features/cases/casesSlice";
-export * from "./features/cases/types";
 export * from "./features/cohort";
 export * from "./features/summary/totalCountsSlice";
 export * from "./features/genomic/ssmsTableSlice";
@@ -45,7 +45,6 @@ export * from "./features/cart";
 export * from "./features/api/cohortApiSlice";
 export * from "./features/bannerNotification";
 export * from "./features/clinicalDataAnalysis";
-export * from "./features/cases/caseSummarySlice";
 export * from "./features/genomic/ssmsSummary/ssmsSummarySlice";
 export * from "./features/genomic/geneSummary/geneSummarySlice";
 export * from "./features/survival/survivalApiSlice";
