+++ packages/portal-proto/src/features/cases/CaseSummary.tsx
@@ -1,4 +1,4 @@
-import { useCaseSummary, useGetAnnotationsQuery } from "@gff/core";
+import { useGetCasesQuery, useGetAnnotationsQuery } from "@gff/core";
 import { LoadingOverlay } from "@mantine/core";
 import { SummaryErrorHeader } from "@/components/Summary/SummaryErrorHeader";
 import { caseSummaryFields } from "./utils";
@@ -18,24 +18,28 @@ export const CaseSummary = ({
   const [shouldScrollToBio, setShouldScrollToBio] = useState(
     bio_id !== undefined,
   );
-  const { data, isFetching } = useCaseSummary({
-    filters: {
-      content: {
-        field: "case_id",
-        value: case_id,
+  const { data, isFetching } = useGetCasesQuery({
+    request: {
+      filters: {
+        content: {
+          field: "case_id",
+          value: case_id,
+        },
+        op: "=",
       },
-      op: "=",
+      fields: caseSummaryFields,
     },
-    fields: caseSummaryFields,
   });
 
   const { data: annotationCountData, isFetching: isAnnotationCallFetching } =
     useGetAnnotationsQuery({
-      filters: {
-        op: "=",
-        content: {
-          field: "annotations.case_id",
-          value: case_id,
+      request: {
+        filters: {
+          op: "=",
+          content: {
+            field: "annotations.case_id",
+            value: case_id,
+          },
         },
       },
     });
@@ -56,7 +60,7 @@ export const CaseSummary = ({
     <>
       {isFetching ||
       isAnnotationCallFetching ||
-      (data && data.case_id !== case_id) ? (
+      (data && data?.hits?.[0]?.case_id !== case_id) ? (
         <LoadingOverlay visible data-testid="loading-spinner" />
       ) : data &&
         Object.keys(data).length > 0 &&
@@ -64,7 +68,7 @@ export const CaseSummary = ({
         <CaseView
           case_id={case_id}
           bio_id={bio_id}
-          data={data}
+          data={data?.hits?.[0]}
           annotationCountData={annotationCountData?.pagination.total}
           isModal={isModal}
           shouldScrollToBio={shouldScrollToBio}
